+++ src/dms-patterns/core/s3.ts
