+++ src/dms-patterns/core/table-mappings/index.ts
@@ -3,7 +3,7 @@ All Rights Reserved.
 */
 
 export * from './base';
-export * from './s3';
+export * from '../s3';
 export * from './selection';
 export * from './table-settings';
 export * from './transformation';
